FEATURE: fixes

+++ src/securityDevices/routers/handlers/delete-all-security-devices.handler.ts
@@ -1,7 +1,10 @@
 import { Request, Response } from 'express';
 import { jwtService } from '../../../auth/adapters/jwt.service';
 import { HttpStatuses } from '../../../core/types/http-statuses';
-import { securityDevicesCollection } from '../../../db/mongo.db';
+import {
+  refreshTokensCollection,
+  securityDevicesCollection,
+} from '../../../db/mongo.db';
 import { ObjectId } from 'mongodb';
 
 export async function deleteAllSecurityDevicesHandler(
@@ -18,9 +21,32 @@ export async function deleteAllSecurityDevicesHandler(
     return res.sendStatus(HttpStatuses.UNAUTHORIZED);
   }
 
-  await securityDevicesCollection.deleteMany({
-    userId: new ObjectId(payload?.userId),
-    deviceId: { $ne: payload.deviceId },
+  // текущий токен должен быть валиден
+  const tokenDoc = await refreshTokensCollection.findOne({
+    value: refreshToken,
   });
+  if (!tokenDoc || !tokenDoc.isValid) {
+    return res.sendStatus(HttpStatuses.UNAUTHORIZED);
+  }
+
+  const userObjectId = new ObjectId(payload.userId);
+  const currentDeviceId = payload.deviceId;
+
+  // 1) удалить ВСЕ другие devices
+  await securityDevicesCollection.deleteMany({
+    userId: userObjectId,
+    deviceId: { $ne: currentDeviceId },
+  });
+
+  // 2) сделать refresh tokens для других devices НЕвалидными
+  await refreshTokensCollection.updateMany(
+    {
+      userId: userObjectId,
+      deviceId: { $ne: currentDeviceId },
+      isValid: true,
+    },
+    { $set: { isValid: false } },
+  );
+
   return res.sendStatus(HttpStatuses.NO_CONTENT);
 }
